Changed the model config and fixed the sync issue

+++ src/models/binding.js
@@ -5,7 +5,7 @@ module.exports = (sequelize, DataTypes) => {
 		'binding',
 		{
 			BindingId: {
-				type: DataTypes.INTEGER(11),
+				type: DataTypes.INTEGER,
 				allowNull: false,
 				primaryKey: true,
 				autoIncrement: true,
@@ -15,7 +15,7 @@ module.exports = (sequelize, DataTypes) => {
 				allowNull: false,
 			},
 			RetailerId: {
-				type: DataTypes.INTEGER(11),
+				type: DataTypes.INTEGER,
 				allowNull: false,
 				references: {
 					model: 'retailer',
@@ -27,7 +27,7 @@ module.exports = (sequelize, DataTypes) => {
 				allowNull: false,
 			},
 			BindingStatusId: {
-				type: DataTypes.INTEGER(11),
+				type: DataTypes.INTEGER,
 				allowNull: false,
 				references: {
 					model: 'bindingstatus',
@@ -45,12 +45,12 @@ module.exports = (sequelize, DataTypes) => {
 			// 	defaultValue: sequelize.literal('CURRENT_TIMESTAMP'),
 			// },
 			CreatedBy: {
-				type: DataTypes.INTEGER(11),
+				type: DataTypes.INTEGER,
 				allowNull: true,
 				defaultValue: '1',
 			},
 			UpdatedBy: {
-				type: DataTypes.INTEGER(11),
+				type: DataTypes.INTEGER,
 				allowNull: true,
 				defaultValue: '1',
 			},
